chnage pck app title logic
chnage pck app title logic
