+++ src/components/Cataloguing/index.js
@@ -2,6 +2,6 @@ export { default as ActionsMenuButton } from './Marc/ActionsMenu';
 export { default as Custom006 } from './Result/Tags/Custom006';
 export { default as Custom007 } from './Result/Tags/Custom007';
 export { default as Custom008 } from './Result/Tags/Custom008';
-export { default as CustomLeader } from './Result/Tags/CustomLeader';
+export { default as MarcLeader } from './Marc/MarcLeader';
 export { default as VariableFields } from './Marc/VariableFields';
 export { default as MarcRecord } from './MarcRecord';
