+++ src/components/Search/Result/components/SearchResultPane.js
@@ -37,6 +37,7 @@ class SearchResultPane extends React.Component<Props, {}> {
         authorityResults,
         isReady,
         handleDetails,
+        translate,
         bibsOnly,
         loading,
         autOnly,
@@ -47,7 +48,7 @@ class SearchResultPane extends React.Component<Props, {}> {
           padContent={(marcJSONRecords.length > 0) || isFetching}
           defaultWidth="fill"
           actionMenu={ActionMenu}
-          paneTitle={<FormattedMessage id="ui-marccat.search.record" />}
+          paneTitle={translate({ id:'ui-marccat.search.record' })}
           paneSub={(mergedRecord && mergedRecord.length > 0) ? message : messageNoContent}
           appIcon={{ app: C.META.ICON_TITLE }}
           firstMenu={firstMenu}
