+++ translations/ui-marccat/en.json
@@ -1,6 +1,5 @@
 {
-  "MARCcat": "MARCcat",
-  "marccat": "MARCcat",
+  "app.title" : "MARCcat",
   "noResult": "No Result found",
   "settings.general": "General",
   "settings.general.message": "These are your general app settings.",
